refactor: Split ConformanceReport.ts into smaller modules
Split the large ConformanceReport component (510 lines) into focused,
maintainable modules:

- types.ts (45 lines): Type definitions and interfaces
- data-processor.ts (63 lines): Data fetching and processing logic
- render-utils.ts (34 lines): Utility functions for rendering
- section-renderers.ts (347 lines): Section rendering functions
- ConformanceReport.ts (88 lines): Main component orchestration

Benefits:
- 83% reduction in main file size (510 → 88 lines)
- Each module has a single, clear responsibility
- Easier to test individual components
- Improved code reusability
- Better maintainability and extensibility

All functionality preserved - web demo builds successfully.

Closes #845
